refactor: 게시물, 필터 상태 관리 최적화, 관련된 타입 및 컴포넌트 수정
