+++ src/entities/post/model/type.ts
@@ -1,3 +1,4 @@
+// src/entities/post/model/type.ts
 interface Post {
   id: number
   userId: number
@@ -6,6 +7,16 @@ interface Post {
   createdAt: string
   updatedAt: string
   body: string
+  author: {
+    username: string
+    image: string
+    id: number
+  }
+  reactions?: {
+    likes: number
+    dislikes: number
+  }
+  tags?: string[]
 }
 
 export default Post
