+++ src/features/comments/api/index.ts
@@ -1,0 +1,3 @@
+export * from "./updateComment"
+export * from "./deleteComment"
+export * from "./likeComment"
