+++ src/features/posts/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./useDeletePost"
+export * from "./useUpdatePost"
