+++ src/features/posts/model/usePostsStore.ts
@@ -8,19 +8,10 @@ interface Post {
   userId?: number
 }
 
-interface Tag {
-  id: string
-  name: string
-  slug: string
-  url: string
-}
-
 interface PostState {
   total: number
   posts: Post[]
   selectedPost: Post | null
-  tags: Tag[]
-  selectedTag: string
   newPost: Post
   showAddDialog: boolean
   showEditDialog: boolean
@@ -29,8 +20,6 @@ interface PostState {
   setTotal: (total: number) => void
   setPosts: (posts: Post[]) => void
   setSelectedPost: (post: Post | null) => void
-  setTags: (tags: Tag[]) => void
-  setSelectedTag: (tag: string) => void
   setNewPost: (post: Post) => void
   setShowAddDialog: (show: boolean) => void
   setShowEditDialog: (show: boolean) => void
@@ -47,8 +36,6 @@ const usePostsStore = create<PostState>((set) => ({
     tags: [],
     userId: 0,
   },
-  tags: [],
-  selectedTag: "",
   newPost: { title: "", body: "", userId: 1 },
   showAddDialog: false,
   showEditDialog: false,
@@ -57,8 +44,6 @@ const usePostsStore = create<PostState>((set) => ({
   setTotal: (total) => set({ total }),
   setPosts: (posts) => set({ posts }),
   setSelectedPost: (post) => set({ selectedPost: post }),
-  setTags: (tags) => set({ tags }),
-  setSelectedTag: (tag) => set({ selectedTag: tag }),
   setNewPost: (post) => set({ newPost: post }),
   setShowAddDialog: (show) => set({ showAddDialog: show }),
   setShowEditDialog: (show) => set({ showEditDialog: show }),
